requirements.txt
Flask==2.3.2
pymongo==4.6.1
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.0.3
numpy==1.24.4
python-dotenv==1.0.0
pytest==7.4.0
